requirements.txt
beautifulsoup4==4.9.3
cycler==0.10.0
fastprogress==1.0.0
feedparser==6.0.2
kiwisolver==1.3.1
mecab-python3==1.0.3
Pillow==8.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
regex==2020.11.13
requests==2.25.1
sgmllib3k==1.0.0
six==1.15.0
wordcloud==1.8.1
